Refactor and update dependencies.
This commit includes the refactoring of tailwind configurations to use presets and common configurations from the @repo/ui package. This change helps to ensure consistency of styles across different parts of the application. Moreover, it also includes updates to several dependencies, improving security and performance of the application. Improving the setup for local development, the development server can now be accessed from any device on the network by default when running npm run dev, this is done to promote mobile phone supporting development.